+++ hwconf/luna/m600/luna_m600_display.h
@@ -23,6 +23,6 @@
 #include "stdint.h"
 
 void luna_canbus_start(void);
-float luna_canbus_get_PAS_torque(void);
+float luna_get_pedal_torque(void);
 float get_encoder_error(void);
 #endif /* APP_LUNA_DISPLAY_CANBUS_H_ */