+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * PAS0-9 is 2 zone.
 	* LINEAR-PAS0-5 (0.05, 0.1, 0.15, 0.2, 0.25, 0.3)
 	* EXPO-PAS6-9(0.405, 0.548, 0.74, 1)
+* PAS pedal torque to motor output is progressive (-3.0/poly).
+	
 ---
 
 ### 6.02